# lumen/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/authentication/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "responses": {
                    "200": {
                        "description": "Retorna um token JWT do usuário",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request."
                    }
                }
            }
        },
        "/sample/{category}/things": {
            "get": {
                "tags": [
                    "yourtag"
                ],
                "operationId": "/sample/category/things",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "The category parameter in path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "criteria",
                        "in": "query",
                        "description": "Some optional other parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/file/read": {
            "post": {
                "tags": [
                    "Read XML file"
                ],
                "operationId": "App\\Http\\Controllers\\FileController::read",
                "responses": {
                    "200": {
                        "description": "Retorna um token JWT do usuário",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request."
                    }
                }
            }
        }
    }
}